add stubs for favorites
