+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AddToFavoritesButton.oss.tsx
@@ -1,0 +1,3 @@
+import {AssetKey} from '../assets/types';
+
+export const AddToFavoritesButton = (_: {assetKey: AssetKey}) => null;
