+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/ComputeGraphData.ts
@@ -1,10 +1,10 @@
 import groupBy from 'lodash/groupBy';
+import {filterAssetSelectionByQuery} from 'shared/asset-selection/filterAssetSelectionByQuery.oss';
 
 import {ComputeGraphDataMessageType} from './ComputeGraphData.types';
 import {GraphData, buildGraphData, toGraphId} from './Utils';
 import {AssetNodeForGraphQueryFragment} from './types/useAssetGraphData.types';
 import {GraphDataState} from './useAssetGraphData';
-import {filterAssetSelectionByQuery} from '../asset-selection/filterAssetSelectionByQuery';
 import {doesFilterArrayMatchValueArray} from '../ui/Filters/doesFilterArrayMatchValueArray';
 
 export function computeGraphData({
