+++ js_modules/dagster-ui/packages/ui-core/src/asset-selection/parseAssetSelectionQuery.ts
@@ -5,14 +5,10 @@ import {
   RecognitionException,
   Recognizer,
 } from 'antlr4ts';
-import {FeatureFlag} from 'shared/app/FeatureFlags.oss';
 import {SupplementaryInformation} from 'shared/asset-graph/useAssetGraphSupplementaryData.oss';
 import {AntlrAssetSelectionVisitor} from 'shared/asset-selection/AntlrAssetSelectionVisitor.oss';
 
-import {featureEnabled} from '../app/Flags';
-import {filterByQuery} from '../app/GraphQueryImpl';
 import {AssetGraphQueryItem} from '../asset-graph/useAssetGraphData';
-import {weakMapMemoize} from '../util/weakMapMemoize';
 import {AssetSelectionLexer} from './generated/AssetSelectionLexer';
 import {AssetSelectionParser} from './generated/AssetSelectionParser';
 
@@ -32,7 +28,7 @@ export class AntlrInputErrorListener implements ANTLRErrorListener<any> {
   }
 }
 
-type AssetSelectionQueryResult = {
+export type AssetSelectionQueryResult = {
   all: AssetGraphQueryItem[];
   focus: AssetGraphQueryItem[];
 };
@@ -68,24 +64,3 @@ export const parseAssetSelectionQuery = (
     return e as Error;
   }
 };
-
-export const filterAssetSelectionByQuery = weakMapMemoize(
-  (
-    all_assets: AssetGraphQueryItem[],
-    query: string,
-    supplementaryData: SupplementaryInformation,
-  ): AssetSelectionQueryResult => {
-    if (query.length === 0) {
-      return {all: all_assets, focus: []};
-    }
-    if (featureEnabled(FeatureFlag.flagSelectionSyntax)) {
-      const result = parseAssetSelectionQuery(all_assets, query, supplementaryData);
-      if (result instanceof Error) {
-        return {all: [], focus: []};
-      }
-      return result;
-    }
-    return filterByQuery(all_assets, query);
-  },
-  {maxEntries: 20},
-);
