+++ js_modules/dagster-ui/packages/ui-core/src/assets/AddToFavoritesMenuItem.oss.tsx
@@ -1,0 +1,3 @@
+import {AssetKey} from './types';
+
+export const AddToFavoritesMenuItem = (_: {assetKey: AssetKey}) => null;
