+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetActionMenu.tsx
@@ -1,5 +1,6 @@
 import {Button, Icon, Menu, MenuItem, Popover, Spinner, Tooltip} from '@dagster-io/ui-components';
-import {useContext, useMemo} from 'react';
+import {memo, useContext, useMemo} from 'react';
+import {AddToFavoritesMenuItem} from 'shared/assets/AddToFavoritesMenuItem.oss';
 
 import {optionsForExecuteButton, useMaterializationAction} from './LaunchAssetExecutionButton';
 import {assetDetailsPathForKey} from './assetDetailsPathForKey';
@@ -14,7 +15,6 @@ import {AssetKeyInput} from '../graphql/types';
 import {MenuLink} from '../ui/MenuLink';
 import {RepoAddress} from '../workspace/types';
 import {workspacePathFromAddress} from '../workspace/workspacePath';
-
 interface Props {
   path: string[];
   definition: AssetTableDefinitionFragment | null;
@@ -22,7 +22,7 @@ interface Props {
   onRefresh?: () => void;
 }
 
-export const AssetActionMenu = (props: Props) => {
+export const AssetActionMenu = memo((props: Props) => {
   const {repoAddress, path, definition, onRefresh} = props;
   const {
     featureContext: {canSeeMaterializeAction},
@@ -71,6 +71,7 @@ export const AssetActionMenu = (props: Props) => {
         content={
           <Menu>
             {executeItem}
+            <AddToFavoritesMenuItem assetKey={{path}} />
 
             <MenuLink
               text="Show in group"
@@ -125,7 +126,7 @@ export const AssetActionMenu = (props: Props) => {
       </Popover>
     </>
   );
-};
+});
 
 export const useExecuteAssetMenuItem = (
   definition: {
