+++ js_modules/dagster-ui/packages/ui-core/src/op-selection/AntlrOpSelection.ts
@@ -3,10 +3,10 @@ import {FeatureFlag} from 'shared/app/FeatureFlags.oss';
 
 import {AntlrOpSelectionVisitor} from './AntlrOpSelectionVisitor';
 import {GraphQueryItem, filterByQuery} from '../app/GraphQueryImpl';
-import {AntlrInputErrorListener} from '../asset-selection/filterAssetSelectionByQuery';
 import {OpSelectionLexer} from './generated/OpSelectionLexer';
 import {OpSelectionParser} from './generated/OpSelectionParser';
 import {featureEnabled} from '../app/Flags';
+import {AntlrInputErrorListener} from '../asset-selection/parseAssetSelectionQuery';
 
 type OpSelectionQueryResult<T extends GraphQueryItem> = {
   all: T[];
